# etcdserver: restructure auth.Store and auth.User