This attempts to decouple password-related functions, which previously
existed both in the Store and User structs, by splitting them out into a
separate interface, PasswordStore.  This means that they can be more
easily swapped out during testing.

This also changes the relevant tests to use mock password functions
instead of the bcrypt-backed implementations; as a result, the tests are
much faster.

Before:
```
        github.com/coreos/etcd/etcdserver/auth          31.495s
        github.com/coreos/etcd/etcdserver/etcdhttp      91.205s
```

After:
```
        github.com/coreos/etcd/etcdserver/auth          1.207s
        github.com/coreos/etcd/etcdserver/etcdhttp      1.207s
```