@jonboulle 2015-10-27
Alternative to #3389