@xiang90 2015-10-27
@jonboulle One disadvantage of this approach is that we expose a public interface just for testing. I am not of a fan of that. But I do not have any better suggestion right now. Seems to be OK.