@jonboulle 2015-10-30
Will merge on green.